requirements.txt
pandas==2.2.2
pendulum
minio
apache-airflow==2.10.2
apache-airflow-providers-postgres
apache-airflow-providers-amazon
apache-airflow-providers-slack
pangres
aenum
astronomer-cosmos